Refactor JRS Config Options
