+++ flink-docs/src/main/java/org/apache/flink/docs/configuration/ConfigOptionsDocGenerator.java
@@ -77,6 +77,8 @@ public class ConfigOptionsDocGenerator {
                 new OptionsClassLocation("flink-runtime", "org.apache.flink.runtime.shuffle"),
                 new OptionsClassLocation("flink-runtime", "org.apache.flink.runtime.jobgraph"),
                 new OptionsClassLocation(
+                        "flink-runtime", "org.apache.flink.runtime.highavailability"),
+                new OptionsClassLocation(
                         "flink-streaming-java", "org.apache.flink.streaming.api.environment"),
                 new OptionsClassLocation("flink-yarn", "org.apache.flink.yarn.configuration"),
                 new OptionsClassLocation(
